refactor(design-core/script): 调整文件位置和文件引用顺序
